Refactor(WIP): Use embedded Req/Res Structs for Ctx interface implementation
